perform most of the requested changes, re-submit for review.
