+++ app/src/main/java/ch/epfl/cs311/wanderwave/ui/screens/LoginScreen.kt
@@ -5,14 +5,14 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.testTag
 import ch.epfl.cs311.wanderwave.navigation.NavigationActions
-import ch.epfl.cs311.wanderwave.ui.components.login.LoginScreenHeader
+import ch.epfl.cs311.wanderwave.ui.components.login.LoginAppLogo
 import ch.epfl.cs311.wanderwave.ui.components.login.SignInButton
 import ch.epfl.cs311.wanderwave.ui.components.login.WelcomeTitle
 
 @Composable
 fun LoginScreen(navigationActions: NavigationActions) {
   Column(modifier = Modifier.testTag("loginScreen")) {
-    LoginScreenHeader(navigationActions, modifier = Modifier.weight(1.5f))
+    LoginAppLogo(modifier = Modifier.weight(1f))
     WelcomeTitle(modifier = Modifier.weight(4f))
     SignInButton(modifier = Modifier.weight(1f)) { navigationActions.signIn() }
   }
